Quick code overhaul
Tokio and hyper-util were added as dependencies, as a result of some newer changes in the hyper crate. The easiest way to get around those changes was to replace "smol::net::TcpListener" with "tokio::net::TcpListener" and then wrapping it in a TokioIo struct from the hyper-util crate.

These dependencies could be removed if instead the needed traits "Read" and "Write" from the hyper crate are manually implemented by us. Whether that's a good idea or not, is something that could be discussed here ;)